requirements.txt
beautifulsoup4==4.11.2
bs4==0.0.1
fastjsonschema==2.19.1
python-dotenv==0.19.2
requests==2.31.0
requests-html==0.10.0
requests-toolbelt==1.0.0
